Revert "Merge branch 'fix_change_sim' into 'develop'"

See merge request !153

+++ app/src/main/java/com/viettel/mbccs/data/source/remote/request/GetListDsLamByTeamIdRequest.java
@@ -5,5 +5,5 @@ package com.viettel.mbccs.data.source.remote.request;
  */
 
 public class GetListDsLamByTeamIdRequest {
-    public int shopId;
+    public Long shopId;
 }
